+++ app/components/home/ModeTime.tsx
@@ -1,0 +1,31 @@
+"use client";
+
+import { useSelector } from "react-redux";
+import { RootState } from "../../redux/store";
+import StopWatch from "./Stopwatch";
+import React from "react";
+import Focus from "./Focus";
+import Clock from "./Clock";
+
+export const ModeTime = () => {
+  const currentTab = useSelector((state: RootState) => state.app.currentTab);
+
+  const formatTime = (seconds: number): string => {
+    const mins = Math.floor(seconds / 60);
+    const secs = seconds % 60;
+    return `${mins.toString().padStart(2, "0")}:${secs
+      .toString()
+      .padStart(2, "0")}`;
+  };
+  return (
+    <React.Fragment>
+      {currentTab === "focusTime" ? (
+        <Focus formatTime={formatTime} />
+      ) : currentTab === "stopwatch" ? (
+        <StopWatch formatTime={formatTime} />
+      ) : (
+        <Clock />
+      )}
+    </React.Fragment>
+  );
+};
